+++ CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to [camunda-bpmn-js](https://github.com/camunda/camunda-bpmn
 
 ___Note:__ Yet to be released changes appear here._
 
+* `FEAT`: expose resource metadata on external-resource popup entries
+
 ## 5.34.0
 
 * `DEPS`: update to `@bpmn-io/properties-panel@3.52.0`